requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
torchmetrics
thop
numpy==1.26.4
torch_audiomentations
pesq
pystoi
tqdm
matplotlib
pandas
wandb
comet_ml
jsonschema
soundfile
hydra-core


black
isort
pre-commit
